Modified all headers

+++ exemplier/compilation/test_optimization.sh
@@ -10,54 +10,37 @@
 #
 # Chapter "Compilation" / Chapitre "Compilation"
 #
-# Copyright (C) 1995-2016 Alain Lebret (alain.lebret@ensicaen.fr)
+# Copyright (C) 1995-2023 Alain Lebret (alain.lebret@ensicaen.fr)
 #
 
-echo "Testing optimization options for a tiny program"
-gcc -O0 small.c -o small_O0
-gcc -O1 small.c -o small_O1
-gcc -O2 small.c -o small_O2
-gcc -O3 small.c -o small_O3
-gcc -Os small.c -o small_Os
+# Script to compile and test C programs with various optimization levels
 
-echo "Testing optimization options for a larger program"
-gcc -O0 bubbles.c -o bubbles_O0
-gcc -O1 bubbles.c -o bubbles_O1
-gcc -O2 bubbles.c -o bubbles_O2
-gcc -O3 bubbles.c -o bubbles_O3
-gcc -Os bubbles.c -o bubbles_Os
+compile_and_test() {
+    local program_name=$1
 
-ls -l
-echo "small_O0"
-time ./small_O0
-read a
-echo "small_O1"
-time ./small_O1
-read a
-echo "small_O2"
-time ./small_O2
-read a
-echo "small_O3"
-time ./small_O3
-read a
-echo "small_Os"
-time ./small_Os
-read a
-echo "bubbles_O0"
-time ./bubbles_O0
-read a
-echo "bubbles_O1"
-time ./bubbles_O1
-read a
-echo "bubbles_O2"
-time ./bubbles_O2
-read a
-echo "bubbles_O3"
-time ./bubbles_O3
-read a
-echo "bubbles_Os"
-time ./bubbles_Os
-read a
+    echo "Testing optimization options for $program_name"
+    for opt in 0 1 2 3 s; do
+        if ! gcc -O$opt "$program_name.c" -o "${program_name}_O$opt"; then
+            echo "Compilation failed for optimization level O$opt."
+            return 1
+        fi
+        echo "${program_name}_O$opt"
+        time ./"${program_name}_O$opt"
+        read -p "Press any key to continue..." -n 1 -s
+    done
+}
+
+# Compile and test 'small' and 'bubbles' programs
+if [ -f "small.c" ] && [ -f "bubbles.c" ]; then
+    compile_and_test "small"
+    compile_and_test "bubbles"
+else
+    echo "Error: Required source files not found."
+    exit 1
+fi
 
 echo "Suppressing unwanted files..."
-rm small_O0 small_O1 small_O2 small_O3 small_Os bubbles_O0 bubbles_O1 bubbles_O2 bubbles_O3 bubbles_Os a.out *.a *.s *.i *.o *.txt
+read -p "Are you sure you want to delete compiled files? (y/n) " confirm
+if [ "$confirm" = "y" ]; then
+    rm small_O0 small_O1 small_O2 small_O3 small_Os bubbles_O0 bubbles_O1 bubbles_O2 bubbles_O3 bubbles_Os a.out *.a *.s *.i *.o *.txt
+fi
